Get STRING database Protein-Protein Interactions

STRING documentation: https://string-db.org/cgi/help?sessionId=baEZCS5u1RdM

Protocol used for downloading STRING files is https

Usage

get_string_ppi(
  aliases,
  score_threshold = 0L,
  organism = 9606L,
  network_type = "full",
  link_data = "combined_only",
  folder = tempdir(),
  version = "latest",
  versions = NULL
)

Arguments

aliases

character, vector with protein/gene symbols/aliases

score_threshold

integer, default: 0, to get all PPI, ranges between [0-1000], 200 for low, 400 for medium and 700 for high/stringent scoring PPI

organism

integer, default: 9606 (Homo Sapiens), see ?goat::load_genesets_go_bioconductor taxid parameter for possible organism taxIDs

network_type

character, default: 'full', else 'physical' for only STRING documented physical interactions

link_data

character, default: 'combined_only', else 'full' or 'detailed', see STRING documentation

folder

character, default: tempdir(), else given folder path for where to download STRING files, converted to .parquet for compression and query efficiency, if tempdir() the temporary directory with the downloaded files are removed after the R session

version

character, default: 'latest', else a version to check availability, e.g. "12.0", if version not available the available versions are printed

versions

NULL, else character vector with versions to choose from with version

Value

dataframe (tibble) with protein-protein interactions (symbols and STRING IDs) and STRING combined score

Examples

if (FALSE) {
get_string_ppi(c("TP53", "EGFR", "BRCA1", "MTOR", "MYC", "SOX2"))
}