Create gene overview through overlapping genelists information by overlapping significant genes

Create gene overview through overlapping genelists information by overlapping significant genes

Usage

run_genelists_overlap(genelists)

Arguments

genelists

UI value/list of tibbles/dataframes

Value

tibble/dataframe with (annotated) genes and p-value/effectsize info for each genelist, concluding with overlapping genelists by significant genes

Examples

run_genelists_overlap(list(
  A = get(load(system.file("extdata", "example_genelist.rda", package = "goatea"))), 
  B = get(load(system.file("extdata", "example_genelist.rda", package = "goatea")))
))
#> # A tibble: 100 × 9
#>     gene symbol  A_efsi A_pval B_efsi B_pval A_perc B_perc genelist_overlap
#>    <int> <chr>    <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <chr>           
#>  1 14969 gene_29   -4.9 0.0783   -4.9 0.0783  -96.7  -96.7 ""              
#>  2 12070 gene_75   -4.9 0.845    -4.9 0.845   -96.7  -96.7 ""              
#>  3 10534 gene_76    4.8 0.0121    4.8 0.0121 2686.  2686.  "A+B"           
#>  4 15042 gene_12    4.8 0.282     4.8 0.282  2686.  2686.  ""              
#>  5 14654 gene_68   -4.8 0.0203   -4.8 0.0203  -96.4  -96.4 "A+B"           
#>  6 12937 gene_98   -4.8 0.791    -4.8 0.791   -96.4  -96.4 ""              
#>  7 17523 gene_13    4.7 0.201     4.7 0.201  2499.  2499.  ""              
#>  8 15743 gene_35    4.7 0.471     4.7 0.471  2499.  2499.  ""              
#>  9 19501 gene_69   -4.7 0.0133   -4.7 0.0133  -96.2  -96.2 "A+B"           
#> 10 19908 gene_65   -4.6 0.306    -4.6 0.306   -95.9  -95.9 ""              
#> # ℹ 90 more rows