Set significant and number of genes

Set significant and number of genes

set_significant_N_genes(
  genelist,
  significance_by = "pvalue_effectsize",
  pvalue_threshold = 0.05,
  effectsize_threshold = 1,
  keep_max_n_genes = FALSE,
  keep_max_n_genes_by = "pvalue"
)

Arguments

genelist

list, loaded genelist with goatea::read_validate_genelist()

significance_by

string, default: 'pvalue_effectsize', else 'pvalue' or 'effectsize' to set gene significance to TRUE/FALSE in 'signif' column

pvalue_threshold

numeric, default: 0.05, to set gene significance based on pvalue

effectsize_threshold

numeric, default: 1, to set gene significance based on effectsize

keep_max_n_genes

boolean, default: TRUE, filter down by pvalue to max n genes allowed by goat (max(goat::goat_nulldistributions$N))

keep_max_n_genes_by

string, default: 'pvalue', else 'effectsize', order genes based on lowest pvalues or highest absolute effect sizes

Value

genelist with added 'signif' column with TRUE/FALSE values

Examples

set_significant_N_genes(
  get(load(system.file("extdata", "example_genelist.rda", package = "goatea")))
)
#>       symbol  gene       pvalue effectsize signif
#> 1    gene_81 18837 5.471520e-06       -4.3   TRUE
#> 2    gene_77 19922 2.608286e-04        2.4   TRUE
#> 3    gene_17 13765 8.416599e-04       -2.2   TRUE
#> 4    gene_31 14822 1.943295e-03        2.5   TRUE
#> 5    gene_38 12392 2.473132e-03        3.3   TRUE
#> 6    gene_96 19338 3.349992e-03       -3.3   TRUE
#> 7    gene_70 17254 3.575055e-03        4.2   TRUE
#> 8    gene_55 12874 3.598888e-03       -1.1   TRUE
#> 9    gene_72 13860 4.419765e-03       -1.1   TRUE
#> 10   gene_82 19040 5.928995e-03        0.3  FALSE
#> 11   gene_89 16292 6.265346e-03        0.6  FALSE
#> 12   gene_86 15947 7.064525e-03        2.2   TRUE
#> 13   gene_94 18968 7.718095e-03       -3.9   TRUE
#> 14    gene_3 18953 1.082897e-02        3.3   TRUE
#> 15   gene_14 17269 1.189872e-02       -2.4   TRUE
#> 16   gene_76 10534 1.208455e-02        4.8   TRUE
#> 17   gene_18 13317 1.288837e-02        1.8   TRUE
#> 18   gene_95 11862 1.323199e-02       -2.1   TRUE
#> 19   gene_69 19501 1.326578e-02       -4.7   TRUE
#> 20   gene_83 13383 1.574589e-02       -1.9   TRUE
#> 21   gene_51 10162 1.867518e-02       -2.9   TRUE
#> 22   gene_68 14654 2.029972e-02       -4.8   TRUE
#> 23    gene_1 16730 2.035625e-02       -0.2  FALSE
#> 24   gene_50 18362 2.061664e-02       -0.3  FALSE
#> 25   gene_91 11558 2.233164e-02       -4.4   TRUE
#> 26   gene_45 18862 2.257831e-02       -0.9  FALSE
#> 27   gene_23 10244 2.264889e-02       -3.3   TRUE
#> 28   gene_28 12997 2.290224e-02       -0.2  FALSE
#> 29   gene_26 19424 2.411721e-02       -1.7   TRUE
#> 30   gene_49 10593 2.579521e-02       -3.5   TRUE
#> 31   gene_84 12088 2.588124e-02        3.7   TRUE
#> 32   gene_92 14555 2.664260e-02        3.4   TRUE
#> 33   gene_36 18297 2.710420e-02       -1.9   TRUE
#> 34   gene_46 15785 3.223701e-02        2.1   TRUE
#> 35   gene_59 14767 3.259735e-02       -0.6  FALSE
#> 36   gene_80 12272 3.440294e-02        0.3  FALSE
#> 37   gene_73 19244 3.440736e-02        3.0   TRUE
#> 38   gene_47 16613 3.609518e-02        1.2   TRUE
#> 39   gene_15 16230 3.759849e-02       -1.4   TRUE
#> 40   gene_58 13126 3.917926e-02       -2.0   TRUE
#> 41   gene_48 15077 3.976684e-02       -2.0   TRUE
#> 42    gene_8 18324 3.980623e-02       -0.8  FALSE
#> 43    gene_9 17464 4.168482e-02       -3.0   TRUE
#> 44  gene_100 17954 4.195725e-02        2.2   TRUE
#> 45   gene_78 15544 4.259890e-02        2.2   TRUE
#> 46   gene_74 11748 4.320323e-02       -0.8  FALSE
#> 47   gene_42 17176 4.455923e-02       -0.2  FALSE
#> 48   gene_19 16034 4.521529e-02       -4.2   TRUE
#> 49   gene_27 10777 4.590041e-02        0.8  FALSE
#> 50   gene_21 15300 4.778884e-02        0.6  FALSE
#> 51   gene_16 10659 5.682229e-02       -2.4  FALSE
#> 52   gene_43 11421 6.259138e-02        2.9  FALSE
#> 53   gene_29 14969 7.830104e-02       -4.9  FALSE
#> 54   gene_41 16436 1.060913e-01       -3.0  FALSE
#> 55   gene_97 14825 1.094756e-01       -4.4  FALSE
#> 56    gene_7 11758 1.251644e-01        0.7  FALSE
#> 57   gene_90 18830 1.694536e-01       -0.9  FALSE
#> 58   gene_87 18554 1.847286e-01       -1.4  FALSE
#> 59   gene_63 11946 1.880133e-01       -0.4  FALSE
#> 60   gene_13 17523 2.011200e-01        4.7  FALSE
#> 61   gene_62 18117 2.103502e-01       -3.2  FALSE
#> 62   gene_39 10470 2.198964e-01        3.7  FALSE
#> 63   gene_66 12067 2.213642e-01        2.9  FALSE
#> 64   gene_60 18372 2.423701e-01       -0.6  FALSE
#> 65   gene_79 10497 2.586512e-01       -1.3  FALSE
#> 66   gene_11 17372 2.814258e-01        1.4  FALSE
#> 67   gene_12 15042 2.818997e-01        4.8  FALSE
#> 68    gene_5 14578 2.952925e-01       -4.3  FALSE
#> 69   gene_57 18378 3.009721e-01        2.0  FALSE
#> 70   gene_65 19908 3.064217e-01       -4.6  FALSE
#> 71   gene_10 11283 3.278389e-01        1.0  FALSE
#> 72   gene_54 19099 4.216015e-01       -4.4  FALSE
#> 73   gene_85 16057 4.343237e-01        1.7  FALSE
#> 74   gene_25 16883 4.440464e-01       -1.7  FALSE
#> 75   gene_71 19051 4.588497e-01       -4.1  FALSE
#> 76   gene_35 15743 4.711346e-01        4.7  FALSE
#> 77   gene_24 15028 4.790531e-01       -4.1  FALSE
#> 78   gene_99 10452 5.002073e-01        4.5  FALSE
#> 79   gene_64 19911 5.390606e-01        1.5  FALSE
#> 80   gene_88 16875 5.583877e-01       -2.3  FALSE
#> 81    gene_4 18586 5.793203e-01        0.4  FALSE
#> 82   gene_33 13741 5.897144e-01        3.1  FALSE
#> 83   gene_61 14779 7.188174e-01       -3.1  FALSE
#> 84   gene_40 16411 7.214974e-01        2.0  FALSE
#> 85   gene_22 11816 7.720894e-01        1.8  FALSE
#> 86   gene_98 12937 7.906323e-01       -4.8  FALSE
#> 87   gene_93 11537 7.991296e-01        0.3  FALSE
#> 88   gene_67 12006 8.029314e-01        3.4  FALSE
#> 89   gene_75 12070 8.452830e-01       -4.9  FALSE
#> 90   gene_20 18738 8.568102e-01        1.2  FALSE
#> 91    gene_6 12680 9.238778e-01        1.9  FALSE
#> 92   gene_30 11678 9.278401e-01       -0.3  FALSE
#> 93   gene_53 19048 9.305928e-01       -2.9  FALSE
#> 94   gene_34 14867 9.330859e-01       -2.3  FALSE
#> 95   gene_32 17438 9.384276e-01       -4.3  FALSE
#> 96   gene_52 16529 9.505959e-01        0.8  FALSE
#> 97   gene_56 14917 9.514949e-01        2.8  FALSE
#> 98   gene_37 10795 9.542250e-01       -1.3  FALSE
#> 99    gene_2 18348 9.589466e-01       -1.1  FALSE
#> 100  gene_44 19156 9.951943e-01        2.6  FALSE