Set significant and number of genes

Set significant and number of genes

set_significant_N_genes(
  genelist,
  significance_by = "pvalue_effectsize",
  pvalue_threshold = 0.05,
  effectsize_threshold = 1,
  keep_max_n_genes = FALSE,
  keep_max_n_genes_by = "pvalue"
)

Arguments

genelist

list, loaded genelist with goatea::read_validate_genelist()

significance_by

string, default: 'pvalue_effectsize', else 'pvalue' or 'effectsize' to set gene significance to TRUE/FALSE in 'signif' column

pvalue_threshold

numeric, default: 0.05, to set gene significance based on pvalue

effectsize_threshold

numeric, default: 1, to set gene significance based on effectsize

keep_max_n_genes

boolean, default: TRUE, filter down by pvalue to max n genes allowed by goat (max(goat::goat_nulldistributions$N))

keep_max_n_genes_by

string, default: 'pvalue', else 'effectsize', order genes based on lowest pvalues or highest absolute effect sizes

Value

genelist with added 'signif' column with TRUE/FALSE values

Examples

set_significant_N_genes(
  get(load(system.file("extdata", "example_genelist.rda", package = "goatea")))
)
#>       symbol  gene       pvalue effectsize signif
#> 1    gene_77 13971 0.0005761071        1.9   TRUE
#> 2    gene_37 17976 0.0007281853        1.7   TRUE
#> 3    gene_72 14353 0.0019763106        1.3   TRUE
#> 4    gene_62 19047 0.0026180818        1.1   TRUE
#> 5    gene_24 18669 0.0054691315        0.2  FALSE
#> 6    gene_80 10203 0.0056068281       -3.1   TRUE
#> 7     gene_3 18654 0.0057399258       -1.9   TRUE
#> 8    gene_31 13749 0.0059652171       -4.6   TRUE
#> 9    gene_93 11933 0.0070328307       -0.4  FALSE
#> 10   gene_32 11591 0.0091781541        2.1   TRUE
#> 11   gene_16 13820 0.0098611411        1.9   TRUE
#> 12   gene_50 12972 0.0110291413       -2.7   TRUE
#> 13   gene_35 11766 0.0114130165       -3.6   TRUE
#> 14    gene_2 15182 0.0134083821        2.6   TRUE
#> 15   gene_83 12069 0.0147880407       -4.9   TRUE
#> 16   gene_14 17454 0.0151459930        4.8   TRUE
#> 17   gene_18 12577 0.0154045383        3.6   TRUE
#> 18   gene_76 17267 0.0156027339       -0.1  FALSE
#> 19   gene_88 18991 0.0171371438       -2.3   TRUE
#> 20   gene_85 18008 0.0175477143        1.8   TRUE
#> 21   gene_87 16970 0.0177418999       -3.9   TRUE
#> 22   gene_10 10245 0.0185564985        3.6   TRUE
#> 23   gene_11 14167 0.0212549415        2.4   TRUE
#> 24   gene_56 11666 0.0217266041       -0.8  FALSE
#> 25   gene_73 18041 0.0256478642       -0.1  FALSE
#> 26   gene_75 11867 0.0280972908        1.0   TRUE
#> 27   gene_68 12566 0.0301592460       -2.0   TRUE
#> 28   gene_70 13223 0.0314160230        0.0  FALSE
#> 29   gene_15 18773 0.0319263353       -0.2  FALSE
#> 30   gene_42 10425 0.0327860830       -1.9   TRUE
#> 31   gene_61 11943 0.0330343748       -1.2   TRUE
#> 32   gene_47 17289 0.0346179819       -0.6  FALSE
#> 33   gene_84 17210 0.0351033714        4.2   TRUE
#> 34   gene_64 15285 0.0361981355       -0.1  FALSE
#> 35   gene_46 18408 0.0367585684        4.5   TRUE
#> 36   gene_81 11493 0.0374524091        1.1   TRUE
#> 37   gene_29 11811 0.0376296315       -1.8   TRUE
#> 38   gene_78 17297 0.0380002699       -3.2   TRUE
#> 39   gene_89 18764 0.0381417346       -0.3  FALSE
#> 40   gene_22 18233 0.0396144277       -2.0   TRUE
#> 41   gene_17 19030 0.0401735036        3.8   TRUE
#> 42   gene_63 12543 0.0401987046        4.1   TRUE
#> 43   gene_59 10264 0.0410447186       -4.6   TRUE
#> 44   gene_51 18795 0.0435994704       -3.3   TRUE
#> 45   gene_79 16395 0.0447389679        2.0   TRUE
#> 46    gene_5 13477 0.0452187020       -2.5   TRUE
#> 47   gene_36 19754 0.0455379509        0.1  FALSE
#> 48   gene_27 12137 0.0457239992       -4.0   TRUE
#> 49   gene_71 19591 0.0472080140       -1.0   TRUE
#> 50   gene_19 11321 0.0492537063        1.1   TRUE
#> 51   gene_33 16692 0.0500722567        0.1  FALSE
#> 52   gene_60 15627 0.0803200770       -3.6  FALSE
#> 53   gene_43 14141 0.0925912534        3.8  FALSE
#> 54   gene_53 15513 0.0960559933       -2.3  FALSE
#> 55    gene_4 18229 0.1001915652        1.3  FALSE
#> 56   gene_38 11429 0.1107157922       -3.0  FALSE
#> 57   gene_45 11023 0.1225016047        5.0  FALSE
#> 58   gene_69 14841 0.1287529941        3.3  FALSE
#> 59   gene_97 10828 0.1346589871        3.8  FALSE
#> 60   gene_39 14632 0.1450617983        4.3  FALSE
#> 61   gene_55 19963 0.1515757250        4.2  FALSE
#> 62   gene_82 10679 0.1675504107        0.3  FALSE
#> 63   gene_48 10454 0.1954471020        1.3  FALSE
#> 64   gene_25 19004 0.2083527534       -1.5  FALSE
#> 65   gene_86 16145 0.2550312885        1.9  FALSE
#> 66   gene_28 13600 0.3233845084       -1.5  FALSE
#> 67   gene_74 18124 0.3353291241        2.4  FALSE
#> 68    gene_8 16454 0.3499692532        2.3  FALSE
#> 69   gene_23 17882 0.3731017712        3.2  FALSE
#> 70   gene_96 12308 0.3745010077        4.8  FALSE
#> 71    gene_9 12282 0.3801313830        0.3  FALSE
#> 72  gene_100 19139 0.3894702496        4.5  FALSE
#> 73   gene_40 19010 0.3950270824        4.3  FALSE
#> 74   gene_95 14263 0.4087141513       -1.2  FALSE
#> 75   gene_12 12763 0.4351471254        5.0  FALSE
#> 76   gene_34 16847 0.4351858142        5.0  FALSE
#> 77   gene_20 12460 0.4433733128       -1.9  FALSE
#> 78    gene_6 11722 0.4875204958        3.8  FALSE
#> 79   gene_30 11078 0.5001130297        2.2  FALSE
#> 80    gene_1 10333 0.5403093128        2.5  FALSE
#> 81   gene_67 18157 0.5504981769       -4.5  FALSE
#> 82   gene_91 14584 0.5618272607       -0.8  FALSE
#> 83   gene_41 17539 0.5714313253        1.5  FALSE
#> 84   gene_21 11309 0.5737756956       -3.9  FALSE
#> 85   gene_58 10991 0.5743754293       -1.0  FALSE
#> 86   gene_26 12698 0.5953526736       -4.1  FALSE
#> 87   gene_49 19790 0.6056210295        0.9  FALSE
#> 88   gene_66 11027 0.6598133688        1.0  FALSE
#> 89   gene_57 19013 0.6670528062       -4.8  FALSE
#> 90   gene_92 11532 0.6761687007        4.7  FALSE
#> 91   gene_54 11182 0.6915581074        1.4  FALSE
#> 92   gene_44 16473 0.7670898249       -0.5  FALSE
#> 93   gene_98 19388 0.7766925020       -2.2  FALSE
#> 94   gene_94 17779 0.8444629515       -4.3  FALSE
#> 95   gene_52 13287 0.9258234095       -3.1  FALSE
#> 96    gene_7 18906 0.9268065962        0.0  FALSE
#> 97   gene_90 15371 0.9688015921        1.8  FALSE
#> 98   gene_99 10629 0.9714059386       -3.7  FALSE
#> 99   gene_65 15882 0.9806994354       -0.7  FALSE
#> 100  gene_13 19140 0.9933005570        0.7  FALSE